refactored sla to package, organized slack package
