+++ .gitignore
@@ -17,3 +17,4 @@ slab
 
 config.toml
 oncall-slab
+dev-slab
