+++ sla/timer.go
@@ -1,14 +1,42 @@
-package slack
+package sla
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/tylerconlee/slab/config"
-	"github.com/tylerconlee/slab/zendesk"
+	"github.com/tylerconlee/slab/slack"
+
+	Zen "github.com/tylerconlee/slab/zendesk"
 )
 
+// RunTimer takes the interval from the config, and at each loop iteration,
+// grabs the latest tickets, checks for upcoming SLAs and send notifications if
+// appropriate
+func RunTimer(interval time.Duration) {
+	log.Info("Starting timer with ", interval, " intervals")
+	t := time.NewTicker(interval)
+	for {
+		active := Zen.CheckSLA()
+		log.Info("Successfully grabbed and parsed tickets from Zendesk")
+		log.Info("Checking ticket notifications...")
+		for _, ticket := range active {
+
+			if ticket.Priority != nil {
+				send, notify := UpdateCache(ticket)
+				if send {
+					n := PrepNotification(ticket, notify)
+					slack.SendSLAMessage(n, ticket)
+				}
+			}
+		}
+		log.Info("Ticket notifications sent. Returning to idle state.")
+		<-t.C
+	}
+}
+
 // PrepNotification takes a given ticket and what notification level and returns a string to be sent to Slack.
-func PrepNotification(ticket zendesk.ActiveTicket, notify int64) (notification string) {
+func PrepNotification(ticket Zen.ActiveTicket, notify int64) (notification string) {
 	log.Debug("Preparing notification for", ticket.ID)
 	c := config.LoadConfig()
 	var t, p string
